REGRESSION(140571): Crash in ScrollingCoordinator::mainThreadScrollingReasons during Frame::createView https://bugs.webkit.org/show_bug.cgi?id=107868
Reviewed by Beth Dakin.

When using fixed layout mode, it's possible to update the scrollbar state of the main FrameView before
it's set on the main frame, which can lead to calling in to ScrollingCoordinator to compute main thread
scrolling regions when m_page->mainFrame()->view() is 0. In this case, we don't have any main
thread scrolling reasons and should just return 0.

* page/scrolling/ScrollingCoordinator.cpp:
(WebCore::ScrollingCoordinator::mainThreadScrollingReasons):

Identifier: 126053@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140732 268f45cc-cd09-0410-ab3c-d52691b4dbfc