Disable Clang's -Wreturn-type-c-linkage for Source/WebCore/bindings/v8/ https://bugs.webkit.org/show_bug.cgi?id=107845
Reviewed by Adam Barth.

A recent version of Clang started warning about code in
Source/WebCore/bindings/v8/npruntime.cpp.

In this case it is warning about an 'extern "c"' function returning a
non-POD type. But because the function has internal linkage, the
warning doesn't really apply.

This change suppresses the warning in the meantime.

Tested by building with Clang r172454.

* WebCore.gyp/WebCore.gyp:

Identifier: 126115@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140800 268f45cc-cd09-0410-ab3c-d52691b4dbfc