IndexedDB: Remove speculative dispatchEvent crash fix now that root cause is addressed https://bugs.webkit.org/show_bug.cgi?id=108131
Reviewed by Tony Chang.

A patch was landed in r140027 to prevent a null pointer crash in
IDBRequest::dispatchEvent and confirm the source of the crash (which
we couldn't repro locally). Following confirmation from user reports,
the root cause was tracked down to a race condition in MessageQueue,
and was fixed in r140483. This patch is no longer needed, and there's
always been an ASSERT in place to catch regressions while debugging.

No new tests - no behavior change.

* Modules/indexeddb/IDBRequest.cpp:
(WebCore::IDBRequest::dispatchEvent): Remove extraneous test.


Identifier: 126624@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@141351 268f45cc-cd09-0410-ab3c-d52691b4dbfc