[Qt] Make it possible to compile WebKit2 Qt related files without access to internal WK2 C++ API https://bugs.webkit.org/show_bug.cgi?id=108472
Reviewed by Andreas Kling.

.:

When linking WebKit2, also link the WebKit2QML module.

* Source/api.pri:

Source/WebKit2:

Add a new module to the qmake build system that represents the part of the WebKit2 Qt integration
that doesn't depend on WebKit2 internals.

Changed qwebnavigationhistory.cpp to not use any internal headers and compile it as part of the
internals-free module.

* Target.pri:
* UIProcess/API/qt/qwebnavigationhistory.cpp:
* WebKit2.pro:
* WebKit2QML.pri: Added.

Tools:

Add a new module to the qmake build system that represents the part of the WebKit2 Qt integration
that doesn't depend on WebKit2 internals.

* qmake/mkspecs/features/webkit_modules.prf:
* qmake/mkspecs/modules/webkit2qml.prf: Copied from Source/WebKit2/WebKit2.pro.

Identifier: 126700@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@141430 268f45cc-cd09-0410-ab3c-d52691b4dbfc