Use TouchEvent.targetTouches rather than TouchEvent.touches since the order of Touches in TouchList is not guaranteed. https://bugs.webkit.org/show_bug.cgi?id=108426
Reviewed by Dimitri Glazkov.

Neat fix for LayoutTest in http://trac.webkit.org/changeset/141054.

In this LayoutTest context, the length of touchEvent.touches is 2.
The Touch order in TouchList is not guaranteed. Therefore
touchEvent.touches[0] can be another touch.  Rather, we should use
touchEvent.targetTouches here since its length is 1 in this context.

* fast/dom/shadow/touch-event.html:


Identifier: 126786@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@141520 268f45cc-cd09-0410-ab3c-d52691b4dbfc