[GTK][WK2] MiniBrowser fullscreen signals support https://bugs.webkit.org/show_bug.cgi?id=108005
Patch by Manuel Rego Casasnovas <rego@igalia.com> on 2013-02-01
Reviewed by Carlos Garcia Campos.

Shows a label for 2 seconds using overlay (if GTK >= 3.2.0) and hides
the toolbar when entering fullscreen. When leaving fullscreen the
label is hidden (if it's not hidden yet) and the toolbar is shown again.

* MiniBrowser/gtk/BrowserWindow.c:
(_BrowserWindow):
(fullScreenMessageTimeoutCallback):
(webViewEnterFullScreen):
(webViewLeaveFullScreen):
(browserWindowFinalize):
(browserWindowConstructed):

Identifier: 126850@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@141585 268f45cc-cd09-0410-ab3c-d52691b4dbfc