Fix test scripts for EWS bots to get them running once and for all. https://bugs.webkit.org/show_bug.cgi?id=108422
Reviewed by Timothy Horton.

Disabling --quiet option on Windows for now because that fails when we try to use /dev/null.
runtests.py was not passing in a configuration to run-webkit-tests which causes DRT to build in release by default.
However, we build in Debug only on the EWS Windows bots, which caused the build to fail.

* Scripts/webkitpy/tool/steps/runtests.py:
(RunTests.options):
(RunTests.run):


Identifier: 126890@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@141628 268f45cc-cd09-0410-ab3c-d52691b4dbfc