display:none file upload button crashes https://bugs.webkit.org/show_bug.cgi?id=109102
Reviewed by Levi Weintraub.

Source/WebCore:

Test: fast/forms/file/display-none-upload-button.html

* rendering/RenderFileUploadControl.cpp:
(WebCore::nodeWidth):
(WebCore::RenderFileUploadControl::paintObject):
Having an upload button doesn't mean we have a rendered upload button.
Null check the renderer before trying to access it.

LayoutTests:

* fast/forms/file/display-none-upload-button-expected.txt: Added.
* fast/forms/file/display-none-upload-button.html: Added.
Tests that we don't crash. Also exposes a bug that the baseline and height of
the input don't include the height of the filename text.

Identifier: 127302@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@142054 268f45cc-cd09-0410-ab3c-d52691b4dbfc