IndexedDB: De-flake open-during-transaction layout test https://bugs.webkit.org/show_bug.cgi?id=109072
Reviewed by Tony Chang.

This test was observed to be flaky in local runs; sometimes the transaction
would terminate after the third open() call rather than the second, resulting
in a TEXT difference. Added code to keep the transaction alive until all of
the open() calls are complete, and changed expectations to match.

* storage/indexeddb/open-during-transaction-expected.txt:
* storage/indexeddb/resources/open-during-transaction.js:


Identifier: 127565@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@142324 268f45cc-cd09-0410-ab3c-d52691b4dbfc